+++ world/world.h
@@ -16,7 +16,7 @@ public:
     bool collides(const Vec<float>& position) const;
     GameObject* create_player(const Level& level);
     void update(float dt);
-    void move_to(Vec<float>& position, const Vec<float>& size, Vec<float>& velocity);
+    void move_to(Vec<float>& position, const Vec<int>& size, Vec<float>& velocity);
     void load_level(const Level& level);
 
     Tilemap tilemap;
